+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [feat: (#39) variadic function arguments](https://github.com/amp-lang/amp/issues/39)
 - [feat: (#41) type aliases](https://github.com/amp-lang/amp/issues/41)
 - [feat!: (#52) unsized types](https://github.com/amp-lang/amp/issues/52)
+- [feat!: (#34) export struct fields](https://github.com/amp-lang/amp/issues/34)
 
 ### Library
 - feat: implement standard library, I guess